% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{beEst}
\alias{beEst}
\title{Bernoulli estimator}
\usage{
beEst(fileName)
}
\arguments{
\item{fileName:}{the name of a .txt file containig the data that we are going to estimate with}
}
\value{
prints a random number with estimated parameters and plots the distribution using ggplot2
}
\description{
a method to estimate parameter of the Bernoulli distribution , and then produces 
another number following this distribution and then plots them.
}
