% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{biEst}
\alias{biEst}
\title{Binomial estimator}
\usage{
biEst(fileName)
}
\arguments{
\item{fileName:}{the name of a .txt file containig the data that we are going to estimate with}
}
\value{
prints the parameter of this distribution and then
 prints a random number with estimated parameters and plots the distribution using ggplot2
}
\description{
a method to estimate parameter of the Binomial distribution , and then produces 
another number following this distribution and then plots them.
}
