% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bigen}
\alias{bigen}
\title{Binomial random generator}
\usage{
bigen(n, p)
}
\arguments{
\item{n:}{a positive integer as the parameter of Binomial random variable, used for producing a random number between 0 and n}

\item{p:}{a float number as the parameter of Binomial random variable, used for generating Bernouli random variables with parameter p}
}
\value{
a random number from Binomial distribution where the generated Binomial variable will be of (n,p)
}
\description{
a method to create Binomial random variables
}
