% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{bigenPlot}
\alias{bigenPlot}
\title{This function plots a distribution cantainig n Binomial random variables with parameters n,p using ggplot2}
\usage{
bigenPlot(simulations, n, p)
}
\arguments{
\item{simulations:}{number of random numbers produced using bigen with parameters n,p}

\item{n:}{the parameter of Binomial random variable, used for producing random number between 0 and n}

\item{p:}{the parameter of Binomial random variable, used for generating Bernouli random variables with parameter p}
}
\value{
a plot of produced random numbers
}
\description{
This function plots a distribution cantainig n Binomial random variables with parameters n,p using ggplot2
}
