% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernouli.R
\name{brgen}
\alias{brgen}
\title{Bernouli random generator}
\usage{
brgen(p)
}
\arguments{
\item{p:}{a float number as the parameter of the Bernouli distribution}
}
\value{
a random number from Bernouli distribution (0 or 1) where the generated Bernouli variable will be of parameter p.
}
\description{
a method to create Bernouli random variables
}
