% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernouli.R
\name{brgenPlot}
\alias{brgenPlot}
\title{This function plots a distribution containig n Bernouli random variables with parameter p using ggplot2}
\usage{
brgenPlot(simulations, p)
}
\arguments{
\item{simulations:}{number of random numbers produced using brgen with parameter p}

\item{p:}{the parameter of Bernouli random variable (between 0 and 1)}
}
\value{
a plot of produced random numbers
}
\description{
This function plots a distribution containig n Bernouli random variables with parameter p using ggplot2
}
