% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uniform.R
\name{cugen}
\alias{cugen}
\title{Uniform random variable generator}
\usage{
cugen(n)
}
\arguments{
\item{n:}{a numeric as count of uniform random variables}
}
\value{
a random number from uniform distribution where the generated uniform variable will be of parameter (0, 1)
}
\description{
a method to create Uniform random variables
}
