% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uniform.R
\name{dugen}
\alias{dugen}
\title{Uniform random variable generator}
\usage{
dugen(n, min = 0, max = 1)
}
\arguments{
\item{n:}{a numeric as count of uniform random variables}

\item{min:}{a numeric as minimum of uniform distribution}

\item{max:}{a numeric as minimum of uniform distribution}
}
\value{
a random number from uniform distribution where the generated uniform variable will be of parameter (min, max).
}
\description{
a method to create Uniform random variables
}
