% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp (2).R
\name{expgen}
\alias{expgen}
\title{Exponential random generator}
\usage{
expgen(landa, number = 1)
}
\arguments{
\item{landa:}{a float number as the parameter(rate) of desired exponential distribution}

\item{number:}{the number of exponentials needed}
}
\value{
a vector of random numbers from Exponential distribution with parameter landa.
}
\description{
a method to create Exponential random variables
}
