% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp (2).R
\name{expplot}
\alias{expplot}
\title{This function plots a distribution cantainig n exponential random variables with parameters landa  using ggplot2}
\usage{
expplot(landa, number = 1)
}
\arguments{
\item{landa:}{the parameter of exponential random variable}

\item{numbers:}{number of random numbers produced using expgen with parameter landa}
}
\value{
a plot of produced random numbers
}
\description{
This function plots a distribution cantainig n exponential random variables with parameters landa  using ggplot2
}
