% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gagen}
\alias{gagen}
\title{Gamma random generator}
\usage{
gagen(number, k, landa)
}
\arguments{
\item{number:}{the number of random Gamma numbers we need}

\item{landa:}{a float number as the parameter of the underlying exponential distribution}

\item{k:}{the number of exponentials needed}
}
\value{
a vector of  random numbers following Gamma distribution with parameters k and landa.
}
\description{
a method to create Gamma random variables
}
