% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gaplot}
\alias{gaplot}
\title{This function plots a distribution cantainig n gamma random variables with parameters k and landa  using ggplot2}
\usage{
gaplot(number, k, landa)
}
\arguments{
\item{numbers:}{number of random numbers produced using gagen with parameters k and landa}

\item{landa:}{the parameter of underlying exponential random variable}

\item{k:}{the number of exponencials needed}
}
\value{
a plot of produced random numbers
}
\description{
This function plots a distribution cantainig n gamma random variables with parameters k and landa  using ggplot2
}
