% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometric.R
\name{gegen}
\alias{gegen}
\title{creates a Geometric random variable}
\usage{
gegen(p)
}
\arguments{
\item{p:}{the parameter of geometric random variable which is the probability of being successful}
}
\value{
a random number following Geometric distribution with parameter p
}
\description{
creates a Geometric random variable
}
