% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometric.R
\name{geometric}
\alias{geometric}
\title{a method to create geometric random variables}
\usage{
geometric(p)
}
\arguments{
\item{p:}{the parameter of desired geometric random variable}
}
\value{
a random number following Geometric distribution with parameter p which is equal to floor(ln(u)/ln(1-p) )
where u is a unifor number between 0 and 1
}
\description{
a method to create geometric random variables
}
