% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{nogen}
\alias{nogen}
\title{Normal random generator}
\usage{
nogen(u, s)
}
\arguments{
\item{u:}{a float number as the parameter of Normal random variable, determining mean}

\item{s:}{a positive float number as the parameter of Normal random variable, determining variance}
}
\value{
a random number from Normal distribution where the generated Normal variable will be of (u,s)
}
\description{
a method to create Normal random variables
}
