% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{nogenPlot}
\alias{nogenPlot}
\title{This function plots a Normal random variables with parameters u,s using ggplot2}
\usage{
nogenPlot(simulations = 10, u = 0, s = 1)
}
\arguments{
\item{simulations:}{number of random numbers produced using nogen with parameters u,s}

\item{u:}{the parameter of Normal random variable, as the mean of Normal RV}

\item{s:}{the parameter of Normal random variable, as the variance of Normal RV}
}
\value{
a plot of produced random numbers
}
\description{
This function plots a Normal random variables with parameters u,s using ggplot2
}
