% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometric.R
\name{plotGeo}
\alias{plotGeo}
\title{This function plots a distribution cantainig n geometric random variables with parameter p using ggplot2}
\usage{
plotGeo(numbers, p)
}
\arguments{
\item{p:}{the parameter of desired geometric random variable}

\item{numbers:}{number of random numbers produced using gegen with parameter p}
}
\value{
a plot of produced random numbers
}
\description{
This function plots a distribution cantainig n geometric random variables with parameter p using ggplot2
}
