% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Possion.R
\name{plotPo}
\alias{plotPo}
\title{This function plots a distribution cantainig n Possian random variables with parameter lambda*t using ggplot2}
\usage{
plotPo(numbers, lambda, t)
}
\arguments{
\item{numbers:}{number of random numbers produced using pogen with parameter p}

\item{lambda:}{the parameter of underlying exponential random variable}

\item{t:}{the length of time interval}
}
\value{
a plot of produced random numbers
}
\description{
This function plots a distribution cantainig n Possian random variables with parameter lambda*t using ggplot2
}
