% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Possion.R
\name{pogen}
\alias{pogen}
\title{Poisson random generator}
\usage{
pogen(lambda, t)
}
\arguments{
\item{lambda:}{a float number as the parameter of the underlying exponential distribution}

\item{t:}{a float as the length of time interval}
}
\value{
a random number from Poisson distribution where the generated Poisson variable will be of parameter lambda*t.
}
\description{
a method to create Possion random variables
}
