% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uniform.R
\name{rgenerator}
\alias{rgenerator}
\title{Merssene Twister Random number generator}
\usage{
rgenerator(seed, numOfReapeat)
}
\arguments{
\item{seed:}{a numeric input as the seed of random generator}

\item{numOfReapeat:}{a numeric for count of generated random numbers}
}
\value{
a 32 bit unsinged int random number.
}
\description{
a method to generate 32 bit unsinged int random number
}
